tests/build: updated and run tests, added pytest as dependency, removed deprecated tests
