+++ pyproject.toml
@@ -27,6 +27,7 @@ tqdm = "^4.67.1"
 jupyter = "^1.1.1"
 jupyter-core = "5.1.2"
 pre-commit = "^3.8.0"
+pytest = "^8.3.4"
 
 [build-system]
 requires = ["poetry-core"]
